hugai-hub/backend/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
langserve[all]>=0.2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
chromadb>=0.5.0
duckduckgo-search>=6.0.0
pyyaml>=6.0
